+++ examples/Default/StreamResponse/src/app.ts
@@ -1,11 +1,10 @@
 import express from 'express';
-import { YtdlCore, YTDL_NodejsStreamType } from '@ybd-project/ytdl-core';
+import { YtdlCore, toPipeableStream } from '@ybd-project/ytdl-core';
 
 const app = express(),
     ytdl = new YtdlCore({
         hl: 'en',
         gl: 'US',
-        streamType: 'nodejs',
     });
 
 app.get('/api/download', (req, res) => {
@@ -14,11 +13,11 @@ app.get('/api/download', (req, res) => {
     res.setHeader('Content-Type', 'video/mp4');
     res.setHeader('Content-Disposition', 'attachment; filename="video.mp4"');
 
-    ytdl.download<YTDL_NodejsStreamType>(`https://www.youtube.com/watch?v=${VIDEO_ID}`, {
+    ytdl.download(`https://www.youtube.com/watch?v=${VIDEO_ID}`, {
         filter: 'audioandvideo',
     })
         .then((stream) => {
-            stream.pipe(res);
+            toPipeableStream(stream).pipe(res);
         })
         .catch((err) => {
             res.json({
@@ -33,11 +32,11 @@ app.get('/api/download/audio', (req, res) => {
     res.setHeader('Content-Type', 'audio/mp3');
     res.setHeader('Content-Disposition', 'attachment; filename="audio.mp3"');
 
-    ytdl.download<YTDL_NodejsStreamType>(`https://www.youtube.com/watch?v=${VIDEO_ID}`, {
+    ytdl.download(`https://www.youtube.com/watch?v=${VIDEO_ID}`, {
         filter: 'audioonly',
     })
         .then((stream) => {
-            stream.pipe(res);
+            toPipeableStream(stream).pipe(res);
         })
         .catch((err) => {
             res.json({
@@ -52,11 +51,11 @@ app.get('/api/download', (req, res) => {
     res.setHeader('Content-Type', 'video/mp4');
     res.setHeader('Content-Disposition', 'attachment; filename="video.mp4"');
 
-    ytdl.download<YTDL_NodejsStreamType>(`https://www.youtube.com/watch?v=${VIDEO_ID}`, {
+    ytdl.download(`https://www.youtube.com/watch?v=${VIDEO_ID}`, {
         filter: 'videoonly',
     })
         .then((stream) => {
-            stream.pipe(res);
+            toPipeableStream(stream).pipe(res);
         })
         .catch((err) => {
             res.json({