+++ src/core/clients/Base.ts
@@ -1,11 +1,16 @@
+type RequestOptions = { payload: string; headers: Record<string, any> };
+
 import type { YT_PlayerApiResponse, YTDL_InnertubeResponseInfo, YTDL_RequestOptions } from '@/types';
+import type { UpperCaseClientTypes } from '@/types/_internal';
 
 import { Platform } from '@/platforms/Platform';
 
 import { PlayerRequestError, UnrecoverableError } from '@/core/errors';
 import { Fetcher } from '@/core/Fetcher';
 
-import type { YTDL_ClientsParams } from '@/utils/Clients';
+import { Logger } from '@/utils/Log';
+
+import type { ClientsParams } from './meta/Clients';
 
 const SHIM = Platform.getShim();
 
@@ -29,11 +34,11 @@ export default class Base {
         return null;
     }
 
-    static request<T = YT_PlayerApiResponse>(url: string, requestOptions: { payload: string; headers: Record<string, any> }, params: YTDL_ClientsParams): Promise<YTDL_InnertubeResponseInfo<T>> {
+    static request<T = YT_PlayerApiResponse>(url: string, requestOptions: RequestOptions, params: ClientsParams, clientName: UpperCaseClientTypes | 'Next'): Promise<YTDL_InnertubeResponseInfo<T>> {
         return new Promise(async (resolve, reject) => {
             const HEADERS: Record<string, string> = {
                     'Content-Type': 'application/json',
-                    'X-Goog-Visitor-Id': params.options.visitorData || '',
+                    'X-Origin': 'https://www.youtube.com',
                     ...requestOptions.headers,
                 },
                 OPTS: YTDL_RequestOptions = {
@@ -46,11 +51,15 @@ export default class Base {
                     originalProxy: params.options.originalProxy,
                 },
                 IS_NEXT_API = url.includes('/next'),
-                ALLOW_RETRY_REQUEST = OPTS.originalProxy && SHIM.runtime !== 'browser',
-                responseHandler = (response: YT_PlayerApiResponse) => {
+                ALLOW_RETRY_REQUEST = !params.options.disableRetryRequest && (((OPTS.originalProxy || OPTS.rewriteRequest) && SHIM.runtime !== 'browser') || HEADERS['Authorization']),
+                responseHandler = (response: YT_PlayerApiResponse, isRetried = false) => {
                     const PLAY_ERROR = this.playError(response);
 
                     if (PLAY_ERROR) {
+                        if (!isRetried && ALLOW_RETRY_REQUEST) {
+                            return retryRequest(PLAY_ERROR);
+                        }
+
                         return reject({
                             isError: true,
                             error: PLAY_ERROR,
@@ -61,6 +70,10 @@ export default class Base {
                     if (!IS_NEXT_API && (!response.videoDetails || params.videoId !== response.videoDetails.videoId)) {
                         const ERROR = new PlayerRequestError('Malformed response from YouTube', response, null);
                         ERROR.response = response;
+
+                        if (!isRetried && ALLOW_RETRY_REQUEST) {
+                            return retryRequest(ERROR);
+                        }
 
                         return reject({
                             isError: true,
@@ -75,27 +88,41 @@ export default class Base {
                         contents: response as T,
                     });
                 },
-                retryRequest = () => {
+                retryRequest = (error?: Error) => {
                     OPTS.originalProxy = undefined;
-                    Fetcher.request<YT_PlayerApiResponse>(url, OPTS)
-                        .then(responseHandler)
-                        .catch((err) => {
-                            reject({
-                                isError: true,
-                                error: err,
-                                contents: null,
+                    OPTS.rewriteRequest = undefined;
+
+                    const HEADERS = new SHIM.polyfills.Headers(OPTS.requestOptions?.headers);
+
+                    HEADERS.delete('Authorization');
+
+                    if (!OPTS.requestOptions) {
+                        OPTS.requestOptions = {};
+                    }
+                    OPTS.requestOptions.headers = HEADERS;
+
+                    Logger.debug(`[ ${clientName} ]: <info>Wait 2 seconds</info> and <warning>retry request...</warning> (Reason: <error>${error?.message}</error>)`);
+                    setTimeout(() => {
+                        Fetcher.request<YT_PlayerApiResponse>(url, OPTS)
+                            .then((res) => responseHandler(res, true))
+                            .catch((err) => {
+                                reject({
+                                    isError: true,
+                                    error: err,
+                                    contents: null,
+                                });
                             });
-                        });
+                    }, 2000);
 
                     return;
                 };
 
             try {
                 Fetcher.request<YT_PlayerApiResponse>(url, OPTS)
-                    .then(responseHandler)
+                    .then((res) => responseHandler(res, false))
                     .catch((err) => {
                         if (ALLOW_RETRY_REQUEST) {
-                            return retryRequest();
+                            return retryRequest(err);
                         }
 
                         reject({