+++ src/core/clients/TvEmbedded.ts
@@ -1,10 +1,10 @@
-import { Clients, YTDL_ClientsParams } from '@/utils/Clients';
+import { Clients, ClientsParams } from './meta/Clients';
 import Base from './Base';
 
 export default class TvEmbedded {
-    static async getPlayerResponse(params: YTDL_ClientsParams) {
+    static async getPlayerResponse(params: ClientsParams) {
         const { url, payload, headers } = Clients.tvEmbedded(params);
 
-        return await Base.request(url, { payload, headers }, params);
+        return await Base.request(url, { payload, headers }, params, 'TvEmbedded');
     }
 }