+++ src/core/clients/WebEmbedded.ts
@@ -1,10 +1,10 @@
-import { Clients, YTDL_ClientsParams } from '@/utils/Clients';
+import { Clients, ClientsParams } from './meta/Clients';
 import Base from './Base';
 
 export default class WebEmbedded {
-    static async getPlayerResponse(params: YTDL_ClientsParams) {
+    static async getPlayerResponse(params: ClientsParams) {
         const { url, payload, headers } = Clients.webEmbedded(params);
 
-        return await Base.request(url, { payload, headers }, params);
+        return await Base.request(url, { payload, headers }, params, 'WebEmbedded');
     }
 }