+++ src/dependence/Jinter/utils/index.ts
@@ -5,21 +5,21 @@ import Visitor from '../visitor.js';
 export const namedFunction = (name: string, fn: Function) => Object.defineProperty(fn, 'name', { value: name });
 
 export interface JSNode<T extends BaseJSNode> extends BaseJSNode {
-    run(): ReturnType<T['run']>;
+  run(): ReturnType<T['run']>;
 }
 
 export interface JSNodeConstructor<T extends BaseJSNode> {
-    new (node: ESTree.Node, visitor: Visitor): JSNode<T>;
+  new(node: ESTree.Node, visitor: Visitor): JSNode<T>;
 }
 
 export class JinterError extends Error {
-    info?: any;
+  info?: any;
 
-    constructor(message: string, info?: any) {
-        super(message);
+  constructor(message: string, info?: any) {
+    super(message);
 
-        if (info) {
-            this.info = info;
-        }
+    if (info) {
+      this.info = info;
     }
+  }
 }