+++ src/utils/Url.ts
@@ -1,3 +1,6 @@
+import { InternalDownloadOptions } from '@/core/types';
+import { Logger } from './Log';
+
 const BASE_URL = 'https://www.youtube.com',
     URL_REGEX = /^https?:\/\//,
     ID_REGEX = /^[a-zA-Z0-9-_]{11}$/,
@@ -33,8 +36,12 @@ class Url {
         return `${BASE_URL}/tv`;
     }
 
-    static getRefreshTokenApiUrl() {
+    static getTokenApiUrl() {
         return `${BASE_URL}/o/oauth2/token`;
+    }
+
+    static getDeviceCodeApiUrl() {
+        return `${BASE_URL}/o/oauth2/device/code`;
     }
 
     static validateID(id: string): boolean {